Removed unnecessary int cast.

+++ src/com/substanceofcode/gps/GpsPositionParser.java
@@ -302,10 +302,10 @@ public class GpsPositionParser {
         }
 
 
-        // if we have a speed value, work out the Miles Per Hour
+        // if we have a speed value, work out the Km Per Hour
         if (groundSpeed > 0) {
             // km/h = knots * 1.852
-            speed = (int) ((groundSpeed) * 1.852);
+            speed = ((groundSpeed) * 1.852);
 
             if (speed > maxSpeed){
             	maxSpeed=speed;
